# Документацию API (Swagger)/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8090
    description: Generated server url
...
paths:
  /work-flow/id:
    get:
      tags:
        - Рабочая область
      summary: Get a task by its id
      operationId: get
      parameters:
        - name: id
          in: path
          description: id of task to be searched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found the work Flow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkFlowDTO'
        "404":
          description: work Flow not found
        "400":
          description: Invalid id supplied
    put:
      tags:
        - Рабочая область
      operationId: edit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkFlowDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkFlowDTO'
    delete:
      tags:
        - Рабочая область
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkFlowDTO'
...
  /task/id:
    get:
      tags:
        - Mенеджер задач
      summary: Get a task by its id
      operationId: get_2
      parameters:
        - name: id
          in: path
          description: id of task to be searched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "404":
          description: task not found
        "400":
          description: Invalid id supplied
        "200":
          description: Found the task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
    put:
      tags:
        - Mенеджер задач
      operationId: edit_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskDTO'
    delete:
      tags:
        - Mенеджер задач
      operationId: delete_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskDTO'
...
  /work-flow:
    get:
      tags:
        - Рабочая область
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkFlowDTO'
    post:
      tags:
        - Рабочая область
      operationId: add
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkFlowDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkFlowDTO'
...
  /task:
    get:
      tags:
        - Mенеджер задач
      operationId: getAll_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDTO'
    post:
      tags:
        - Mенеджер задач
      operationId: add_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskDTO'
...
components:
  schemas:
    BoardPreset:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        name:
          type: string
        text:
          type: string
        boardType:
          type: string
          enum:
            - BoardType()
            - BoardType()
            - BoardType()
        boardView:
          type: string
          enum:
            - BoardView()
            - BoardView()
        new:
          type: boolean
    Role:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        code:
          maxLength: 255
          minLength: 0
          type: string
        name:
          maxLength: 255
          minLength: 0
          type: string
        active:
          type: boolean
        new:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        userName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        active:
          type: boolean
        password:
          type: string
        activationCode:
          type: string
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Role'
        admin:
          type: boolean
        new:
          type: boolean
    WorkFlowDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
        boardPreset:
          type: array
          items:
            $ref: '#/components/schemas/BoardPreset'
...
    Contractor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        type:
          type: string
          enum:
            - ContractorType()
            - ContractorType()
            - ContractorType()
            - ContractorType()
        email:
          type: array
          items:
            $ref: '#/components/schemas/Email'
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        contractorAddress:
          type: array
          items:
            $ref: '#/components/schemas/ContractorAddress'
        new:
          type: boolean
    ContractorAddress:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        description:
          type: string
        address:
          type: string
        fias:
          type: string
        phone:
          type: string
        contactPersone:
          type: string
        new:
          type: boolean
    Email:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        text:
          type: string
        active:
          type: boolean
        new:
          type: boolean
    Notification:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        title:
          type: string
        text:
          type: string
        contractor:
          $ref: '#/components/schemas/Contractor'
        new:
          type: boolean
    Phone:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        version:
          type: integer
          format: int32
        code:
          type: string
        text:
          type: string
        active:
          type: boolean
        new:
          type: boolean
    TaskDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        title:
          type: string
        text:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
...